refactor(chess2): move all util classes to utils2
